refactor: cleanup dashboard and document related feature and page files
