--- src/features/dashboard-page/components/index.ts
@@ -1,4 +1,0 @@
-export { default as PageHeader } from "./PageHeader";
-export { default as RecentDocumentsSection } from "./RecentDocumentsSection";
-export { default as DashboardSidebar } from "./Sidebar";
-export { default as SuggestedTemplatesSection } from "./SuggestedTemplatesSection";
