+++ src/features/dashboard/components/SuggestedTemplatesSection.tsx
