+++ src/features/dashboard/index.ts
@@ -1,0 +1,3 @@
+export { default as RecentDocumentsSection } from "./components/RecentDocumentsSection";
+export { default as DashboardSidebar } from "./components/Sidebar";
+export { default as SuggestedTemplatesSection } from "./components/SuggestedTemplatesSection";
