--- src/features/document-listing/components/index.ts
@@ -1,1 +1,0 @@
-export { default as DocumentList } from "./DocumentList";
