--- src/features/documents-page/components/index.ts
@@ -1,2 +1,0 @@
-export { default as DocumentSearch } from "./DocumentSearch";
-export { default as DocumentsSection } from "./DocumentsSection";
