+++ src/features/documents/index.ts
@@ -1,0 +1,5 @@
+export { default as DocumentList } from "./components/DocumentList";
+export { default as DocumentSearch } from "./components/DocumentSearch";
+export { default as MyDocumentsSection } from "./components/MyDocumentsSection";
+
+export { default as useDocumentListQuery } from "./hooks/useDocumentListQuery";
