+++ src/shared/components/common/PageHeader.tsx
@@ -6,11 +6,17 @@ interface PageHeaderProps {
   renderAction?: () => React.ReactNode;
 }
 
-const PageHeader = ({ title, description, renderAction }: PageHeaderProps) => {
+export const PageHeader = ({
+  title,
+  description,
+  renderAction,
+}: PageHeaderProps) => {
   return (
     <header className="flex justify-between flex-wrap gap-5 items-center">
       <div className="flex flex-col gap-1">
-        <h1 className="md:text-2xl text-lg font-bold tracking-tight">{title}</h1>
+        <h1 className="md:text-2xl text-lg font-bold tracking-tight">
+          {title}
+        </h1>
         <p className="text-xs md:text-sm text-gray-600 dark:text-gray-400">
           {description}
         </p>
@@ -19,5 +25,3 @@ const PageHeader = ({ title, description, renderAction }: PageHeaderProps) => {
     </header>
   );
 };
-
-export default PageHeader;
